+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Changed
 - Make the search case insensitive
+- Move the logic to create html buttons in the ui class
 
 ----